V0.1 (#7)
* separate core from cli + start doc

* pass lib from .so to .a

* add custom initialiser feature

* check rd perf + clean wrap cereal dep

* start implementing cereal serialization

* Improve post process package + check monod model

* improve way to get particle properties with kokkos_view in post_process and data_exporter

* Refractor kernel into functor + reduce branching in move process + check cstr simulation

* Refractor and add autogenerated compile_time flag + start impl probes

Remove all compile time flag from core and lib to put them into Meson
Generated header + move model selection output dir ("autogenerated").
Working poc implementation of probes to export some events during
simulation using fixed size buffer.

Check RTD for 0D reactor : OK with same stochastic algo

* Fix(simulation) RTD for batch and continuous reactor; Fix(init) gas scalar initialiaser

* Fix(model): mass_transfer + feed support, optimisation(MC), refractor(data_exporter_wip)

MC: use of Kokkos Scatter to limit use of atomic with host space +
change algortihm to remove dead particle

* WIP(Data exporter): new data exporter architecture

* feat: Implement DataExporter and MPI Feed

- Added DataExporter feature to facilitate data export.
- Split DataExporter into two components: main file for host node and partial file for worker nodes.
- Workers now write particle data efficiently.
- Fixed Kokkos/MPI thread count settings for improved performance.
- Updated .dockerfile to ensure proper environment setup.

* feat,fix: add mass calculation function and improve Monod uptake

fix: dataexporter OK
fix: correct weight calculation
feat: start updating postprocess package to handle new export format

* chore, fix(dataexport): refractor, new files

chore: move simulation parameter to core
fix: datexporter file path + integration in to whole algorithm to
correclty export data
fix: post process data with multiple files

* fix: matrix layout (row/column major) between Eigen and Kokkos

* chore + fix rtd

* misc: test sanofi ok

* fix(monod model): fix model to have correct final biomass concentration

* feat: steady_state test

* feat: test model parameters

* Fix Monod and Uptake model, working in steady state

* script steady

* perf: improve flowmap creation and transition performance

* add external hdf5 compressor

* opti: bench and add profiling

* feat: add prng to division model function

* opti,feat: bench try opti + start impl sunya's model

* refractor: model_uptake rename k_pts

* feat(model) fix uptake, implem metabolism

* fix(simulation): fix oxygen transfer and henry constant

* fix(cli): scalar gas-liquid from files, todo find better fix

* fix(main): confirm overriding when using MPI

* chore

* feat,fix,,chore: working 0.0.9 version

* version upgrade

---------

+++ apps/libs/common/public/common/execinfo.hpp
@@ -3,14 +3,15 @@
 
 #include <cstddef>
 #include <iostream>
+#include <cstdint> 
 
 struct ExecInfo
 {
-  size_t n_rank;
-  size_t current_rank;
-  size_t thread_per_process;
+  uint32_t n_rank;
+  uint32_t current_rank;
+  uint32_t thread_per_process;
   bool verbose;
-  size_t run_id;
+  uint64_t run_id;
 #if defined(__cpp_lib_hardware_interference_size)
   // default cacheline size from runtime
   static constexpr size_t cache_line_size =
